refactor: satisfy Phase 1 modularity guard
