+++ scripts/ci/z620-gates-command-coverage.json
@@ -1,0 +1,18 @@
+{
+  "commandCoverage": {
+    "check-architecture-boundaries": [".github/workflows/ci.yml#audit"],
+    "check-db-access": [".github/workflows/ci.yml#audit"],
+    "check-e2e-contracts-base": [".github/workflows/ci.yml#audit"],
+    "coverage-gate": [".github/workflows/ci.yml#unit"],
+    "e2e-gate-pr": [".github/workflows/e2e-pr.yml#e2e-runner"],
+    "e2e-quarantine-budget": [".github/workflows/ci.yml#audit"],
+    "e2e-smoke": [".github/workflows/e2e-pr.yml#e2e-runner"],
+    "lint-production-warnings": [".github/workflows/ci.yml#audit"],
+    "migration-journal-check": [".github/workflows/ci.yml#audit"],
+    "release-gate-tests": [".github/workflows/ci.yml#unit"],
+    "repo-size-check": [".github/workflows/ci.yml#audit"],
+    "security-guard": [".github/workflows/security.yml#pnpm-audit"],
+    "test-ci-contracts": [".github/workflows/ci.yml#audit"],
+    "workspace-i18n": [".github/workflows/ci.yml#static"]
+  }
+}
