+++ scripts/ci/z620-gates-job-commands.json
@@ -1,0 +1,120 @@
+{
+  "jobCommands": {
+    ".github/workflows/ci.yml#audit": [
+      {
+        "commandId": "check-architecture-boundaries",
+        "argv": ["pnpm", "check:architecture-boundaries"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "check-db-access",
+        "argv": ["pnpm", "check:db-access"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "check-e2e-contracts-base",
+        "argv": ["pnpm", "check:e2e-contracts"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "e2e-quarantine-budget",
+        "argv": ["pnpm", "check:e2e-quarantine-budget"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "lint-production-warnings",
+        "argv": ["pnpm", "lint:production-warnings"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "migration-journal-check",
+        "argv": ["pnpm", "db:migrations:check-journal"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "repo-size-check",
+        "argv": ["pnpm", "repo:size:check"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "test-ci-contracts",
+        "argv": ["pnpm", "test:ci:contracts"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      }
+    ],
+    ".github/workflows/ci.yml#static": [
+      {
+        "commandId": "workspace-i18n",
+        "argv": ["pnpm", "-w", "i18n:check"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      }
+    ],
+    ".github/workflows/ci.yml#unit": [
+      {
+        "commandId": "coverage-gate",
+        "argv": ["pnpm", "coverage:gate"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      },
+      {
+        "commandId": "release-gate-tests",
+        "argv": ["pnpm", "test:release-gate"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      }
+    ],
+    ".github/workflows/e2e-pr.yml#e2e-runner": [
+      {
+        "commandId": "e2e-gate-pr",
+        "argv": ["pnpm", "e2e:gate:pr"],
+        "env": {
+          "E2E_DATABASE_URL": "${{ env.DATABASE_URL }}",
+          "E2E_DATABASE_URL_RLS": "${{ env.DATABASE_URL }}",
+          "PW_EVIDENCE_LANE": "pr-gate"
+        },
+        "projects": ["gate-ks-sq", "gate-mk-contract", "gate-mk-mk", "setup-ks", "setup-mk"],
+        "specs": ["e2e/gate", "e2e/setup.state.spec.ts"]
+      },
+      {
+        "commandId": "e2e-smoke",
+        "argv": ["pnpm", "--filter", "@interdomestik/web", "run", "e2e:smoke"],
+        "env": {
+          "E2E_DATABASE_URL": "${{ env.DATABASE_URL }}",
+          "E2E_DATABASE_URL_RLS": "${{ env.DATABASE_URL }}",
+          "PW_EVIDENCE_LANE": "pr-smoke"
+        },
+        "projects": ["ks-sq", "mk-mk"],
+        "specs": []
+      }
+    ],
+    ".github/workflows/security.yml#pnpm-audit": [
+      {
+        "commandId": "security-guard",
+        "argv": ["pnpm", "security:guard"],
+        "env": {},
+        "projects": [],
+        "specs": []
+      }
+    ]
+  }
+}
