+++ scripts/ci/z620-parity-sources.mjs
@@ -1,0 +1,13 @@
+export const paritySourcePaths = [
+  'apps/web/playwright.config.ts',
+  'package.json',
+  'scripts/ci-local-parity.sh',
+  'scripts/ci/z620-gate-command-lib.mjs',
+  'scripts/ci/z620-gates-command-coverage.json',
+  'scripts/ci/z620-gates-command-policy.json',
+  'scripts/ci/z620-gates-job-commands.json',
+  'scripts/ci/z620-gates-lanes.json',
+  'scripts/ci/z620-gates-loader.mjs',
+  'scripts/ci/z620-gates.json',
+  'scripts/run-e2e-lane.mjs',
+];
