+++ scripts/ci/z620-parity.json
@@ -1,19 +1,25 @@
 {
   "version": 1,
   "workflowDigests": {
-    ".github/workflows/ci.yml": "5583dff9892ab2927886ac53a2f13d30905a711e5e3735462f82a2022f509ae7",
+    ".github/workflows/ci.yml": "a0faa0139d25460f6886d9f9addf82e8e89db74334e30e2e82c328fe4ecae8ce",
     ".github/workflows/security.yml": "72e957bba4773deb3212d6db1d16f398c8fed53e538c6a060ca04f07f688195a",
-    ".github/workflows/e2e-pr.yml": "de83578bf89da45d93f0f35731f9a71258167e19df2ba5b7d6d01d8f3e04c55c",
+    ".github/workflows/e2e-pr.yml": "4ec80d9b258cb812f8ad7756d170f1150a4bcf6bb5da30f504a33c076d5cd572",
     ".github/workflows/sonar-main-gate.yml": "626baa9e4bc3df173f297829076a2ec4711f36fac3de26e66d3cd1fdde7cc0ef",
     ".github/workflows/pilot-gate.yml": "4b9ad1e56df586849ad3b02111178322297ab6d56fa730083ebc1e60e8148bc3",
     ".github/workflows/release-candidate.yml": "ec03f5eb63d19bff1e8502257e36d264fed70733ca682f5820cd74af93f99bac",
     ".github/workflows/cd.yml": "b865817353226d4a372357b004f546100c37d9c3690edbd57bfad5192f2410e2"
   },
   "sourceDigests": {
-    "apps/web/playwright.config.ts": "070c797ef70b5176965ef1e045def272d2e1e80b01a0f7ec8808a87ce3e447e4",
+    "apps/web/playwright.config.ts": "994533a71e3a3f9aba3bf04c300ce58465f9cdb934e40d1b6fea244c241b80ec",
     "package.json": "ac5bda506b0b2fd288dbc1f6dd4dc522efb563cbf21133fae70da845e0161b06",
     "scripts/ci-local-parity.sh": "ed415efa1ecd772408959df535e5a70c686b9842c672ce666e4385af60fba8d2",
     "scripts/ci/z620-gate-command-lib.mjs": "ca1d1fa82bd5cdbd2480144b4074f4a8390e3112091a93736aec76d73a058d10",
+    "scripts/ci/z620-gates-command-coverage.json": "28d789304e38682b22868d78d1e011b765a434a2a316e930467c2517ab240500",
+    "scripts/ci/z620-gates-command-policy.json": "c4daed8e25f52ef1161a0cb48cd4a38845d86f19c6304856c51a001b9482df02",
+    "scripts/ci/z620-gates-job-commands.json": "c28f33e6fae75eb28f57a6687d9537a23dcf3a03478bb55dcdace2c4cb48d1df",
+    "scripts/ci/z620-gates-lanes.json": "5469f810a92e01b471a0f278c04c56531c443b40741faf7e7ac5ba42b7e51e8b",
+    "scripts/ci/z620-gates-loader.mjs": "81cf71c5e6b6c8304f7c933ac0f0aaaec8792a0273948af7a1a1ef5816a499bf",
+    "scripts/ci/z620-gates.json": "60bc11e970ca019323f44bd5b1cda9c4f599341574793e3f6c02329059dd4f38",
     "scripts/run-e2e-lane.mjs": "9beb625efc7ffb6cadf88809063ee8b1b15cd5080c1419de5aa888db504cf685"
   },
   "workflows": {
