+++ scripts/ci/z620-runner.mjs
@@ -15,6 +15,7 @@ import {
   writeJson,
 } from './z620-runner-lib.mjs';
 import { validateGateCommandIds } from './z620-gate-command-lib.mjs';
+import { loadZ620Gates } from './z620-gates-loader.mjs';
 import { validateGateCoverage, validateWorkflowDigests } from './z620-parity-lib.mjs';
 import { resolveRunsRoot, resolveStateRoot } from './z620-resource-policy.mjs';
 
@@ -35,7 +36,7 @@ const sha = String(
 const runsRoot = resolveRunsRoot(args['runs-root'], root);
 const stateRoot = resolveStateRoot(args['state-root'], root);
 const parity = JSON.parse(fs.readFileSync(path.join(root, 'scripts/ci/z620-parity.json'), 'utf8'));
-const gates = JSON.parse(fs.readFileSync(path.join(root, 'scripts/ci/z620-gates.json'), 'utf8'));
+const gates = loadZ620Gates(root);
 
 const problems = [
   ...validateParity(root, parity),
